moved StackTraceTools to utils package to clean up the parent lib directory
